Dummy QR Implementation/requirements.txt
pyzbar==0.1.8
qrcode
Pillow==7.2.0
